Add mapping-formula for form fields
This is meant as a replacement for E:PageForms' mapping_template

This is similar to value-formula and options-formula, except that
this only applies to the display of the option in the form. It does
not alter the saved value like value-formula and options-formula does.

Additionally, unlike options-formula, this shows up for all options
types not just ask queries.

Change-Id: I8fae703ed0958d4d2b3921d8d97e1a9d46277c7b