Let $wgPagePropertiesDisableVersionCheck also disable the API
In a wiki farm context, users in hardcoded groups should not be
notified of changes.  Also, the version check API should not be
left open, as arbitrary people on a wiki farm could do a reflected DoS
against the farm and mediawiki.org. In a trust model where anyone
can create wikis and designate sysops, this needs to be closed.

Change-Id: I8df4aa97a95d8ed2358d71464053c84d1282cbf2